Refactor: rename acceptor to voter again.
