+++ protocol/leaderrotation/common.go
